simplified some of the interface
