+++ in_toto/resolver/file_resolver.py
@@ -25,12 +25,51 @@
 
 import os
 import logging
+from pathspec import PathSpec
 
 import securesystemslib.hash
 
-from in_toto.resolver.resolver import Resolver
+from in_toto.resolver.resolver import Resolver, DEFAULT_SCHEME, get_scheme
 
 LOG = logging.getLogger(__name__)
+
+
+def apply_exclude_patterns(names, exclude_patterns=None):
+  """Exclude matched patterns from passed names."""
+  if not exclude_patterns:
+    return names
+
+  included = set(names)
+
+  exclude_patterns = PathSpec.from_lines('gitwildmatch', exclude_patterns)
+
+  for excluded in exclude_patterns.match_files(names):
+    included.discard(excluded)
+
+  return sorted(included)
+
+
+def apply_left_strip(artifact_uri, lstrip_paths=None):
+  """Internal helper function to left strip dictionary keys based on
+  prefixes passed by the user."""
+  scheme = get_scheme(artifact_uri)
+  if scheme != DEFAULT_SCHEME:
+    artifact_uri = artifact_uri[len(scheme) + 1:]
+    scheme += ":"
+  else:
+    scheme = ""
+
+  if lstrip_paths:
+    # If a prefix is passed using the argument --lstrip-paths,
+    # that prefix is left stripped from the uri passed.
+    # Note: if the prefix doesn't include a trailing /, the dictionary key
+    # may include an unexpected /.
+    for prefix in lstrip_paths:
+      if artifact_uri.startswith(prefix):
+        artifact_uri = artifact_uri[len(prefix):]
+        break
+
+  return scheme + artifact_uri
 
 
 class FileResolver(Resolver):
@@ -50,7 +89,7 @@ class FileResolver(Resolver):
       prepend, generic_uri = generic_uri.split(":", 1)
       prepend += ":"
 
-    norm_paths = super().apply_exclude_patterns(
+    norm_paths = apply_exclude_patterns(
         [os.path.normpath(generic_uri)], exclude_patterns)
     if not norm_paths:
       return norm_paths
@@ -73,7 +112,7 @@ class FileResolver(Resolver):
         dirpaths.append(npath)
 
       if exclude_patterns:
-        dirpaths = super().apply_exclude_patterns(dirpaths, exclude_patterns)
+        dirpaths = apply_exclude_patterns(dirpaths, exclude_patterns)
 
       dirs[:] = [os.path.basename(d) for d in dirpaths]
 
@@ -89,7 +128,7 @@ class FileResolver(Resolver):
               .format(norm_filepath))
 
       if exclude_patterns:
-        filepaths = super().apply_exclude_patterns(filepaths, exclude_patterns)
+        filepaths = apply_exclude_patterns(filepaths, exclude_patterns)
 
       for filepath in filepaths:
         normalized_filepath = filepath.replace("\\", "/")
@@ -99,7 +138,7 @@ class FileResolver(Resolver):
 
   @classmethod
   def get_key_from_uri(cls, resolved_uri):
-    return super().apply_left_strip(resolved_uri, cls.lstrip_paths)
+    return apply_left_strip(resolved_uri, cls.lstrip_paths)
 
   @classmethod
   def get_artifact_hashdict(cls, resolved_uri):
